Touch up ddex entrypoint
