+++ packages/ddex/processor/package.json
@@ -20,19 +20,19 @@
     "@audius/sdk": "*",
     "@aws-sdk/client-s3": "3.504.0",
     "@aws-sdk/credential-provider-ini": "3.556.0",
-    "@hono/node-server": "^1.10.0",
-    "better-sqlite3": "^9.5.0",
-    "cheerio": "^1.0.0-rc.12",
-    "commander": "^12.0.0",
-    "decompress": "^4.2.1",
+    "@hono/node-server": "1.10.0",
+    "better-sqlite3": "9.5.0",
+    "cheerio": "1.0.0-rc.12",
+    "commander": "12.0.0",
+    "decompress": "4.2.1",
     "dotenv": "16.3.1",
-    "file-type": "^19.0.0",
-    "hono": "^4.2.5",
+    "file-type": "19.0.0",
+    "hono": "4.2.5",
     "web3": "4.3.0"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "^7.6.9",
-    "@types/decompress": "^4.2.7",
+    "@types/better-sqlite3": "7.6.9",
+    "@types/decompress": "4.2.7",
     "@types/node": "20.11.10",
     "@typescript-eslint/eslint-plugin": "6.17.0",
     "@typescript-eslint/parser": "6.17.0",
@@ -42,6 +42,6 @@
     "prettier": "3.1.1",
     "tsx": "4.7.2",
     "typescript": "5.3.3",
-    "vitest": "^1.5.0"
+    "vitest": "1.5.0"
   }
 }
